UHM-5404 - Adds 2 features to support this ticket. (#303)
UHM-5404 - Adds 2 features to support this ticket.  1. Adds a new page controller that supports passing a patient identifier and type, and a particular set of configurations, to open a patient dashboard for a patient with that identifier.  2.  Adds the ability to add arbitrary extensions via PIH config as an alternative to doing this programmatically in the CALF.